fix: ApplicationTest.기능_테스트가 통과하지 않는 문제 해결
* fix: 잘못된 Randoms 메소드를 호출하고 있는 문제 해결
* fix: 문구 오타 수정
* refactor: 기타 리팩토링

+++ src/main/java/lotto/Controller.java
@@ -44,7 +44,8 @@ public class Controller {
         LotteryResults results = calculator.getTotalResults(receipts);
         long resultAmount = results.getTotalAmount();
         long purchaseAmount = calculatePurchaseAmount(receipts);
-        double earningRate = resultAmount / purchaseAmount * 100;
+
+        double earningRate = ((double)resultAmount / (double)purchaseAmount) * 100;
         out.printResults(results.toList(), earningRate);
     }
